refactor(deploy): extract pull request logic
we want to create a variant, so lets refactor it first
